chore(common): move CompilerEvent, CompilerCallbacks into types 🙀
- make CompilerCallbacks an interface
- concrete implementations for Node (CLI) and for Testing
- change loadFile to take a string|URL 🙀

supports #7964
