--- developer/src/kmc-keyboard/src/compiler/callbacks.ts
@@ -1,11 +1,0 @@
-export interface CompilerEvent {
-  code: number;
-  message: string;
-};
-
-export default interface CompilerCallbacks {
-  loadFile(baseFilename: string, filename: string): Buffer;
-  loadLdmlKeyboardSchema(): Buffer;
-  loadKvksJsonSchema(): Buffer;
-  reportMessage(event: CompilerEvent): void;
-};
