+++ developer/src/kmc-keyboard/src/main.ts
@@ -2,9 +2,8 @@
 export { default as Compiler } from './compiler/compiler.js';
 export { KeymanWebCompiler } from './compiler/keymanweb-compiler.js';
 export { default as VisualKeyboardCompiler } from './compiler/visual-keyboard-compiler.js';
-export { CompilerEvent, default as CompilerCallbacks } from './compiler/callbacks.js';
 export { default as CompilerOptions } from './compiler/compiler-options.js';
 export { CompilerMessages, CompilerErrorSeverity } from './compiler/messages.js';
 export { default as KMXPlusMetadataCompiler } from './compiler/metadata-compiler.js';
 
-export { KMXBuilder } from '@keymanapp/common-types';
+export { KMXBuilder } from "@keymanapp/common-types";
